Move CacheableGenerator.h -> Generators.h
