+++ test/CMakeLists.txt
@@ -405,7 +405,7 @@ addLinkAndDiscoverTest(CopyableSynchronizationTest)
 
 addLinkAndDiscoverTest(LazyJsonParserTest)
 
-addLinkAndDiscoverTest(CacheableGeneratorTest)
+addLinkAndDiscoverTest(GeneratorsTest)
 
 addLinkAndDiscoverTest(FilterTest engine)
